Add FMIActuatorPluginKnownInputTest
For now, this test just simulates the pendulum + FMU transmission
with a constant input, and saves the result in .mat file to simplify the inspection.

It is already useful on its own for developers, but the idea
is that it can be extended to more inputs (such as as sine) and to
check the behaviour of simple models + transmission agaist their nominal behavior.