First version of the Single Body Fluid Dynamics FMI plugin
This plugin is useful for co-simulating the interaction between a (single) link
in the model and a surrounding medium. It can be used for simple lumped parameters
models, or eventually also for co-simulation with CFD solvers, but the latter case
has not been tested and it would probably be extremly slow.