Fix ComputeJointAcceleration utility function when used with joints attached to "world"
In Gazebo, the "world" link is not a proper link, even if it can  be used in SDF.
For this reason, asking for the API to the "world" link results in a null value.
Before this fix, the plugin that  used ComputeJointAcceleration function were crashing
when attached to a joint attached to the world. With the provided fix, the correct
acceleration for the world (zero) is used. Test have been added to make sure that the problem is
actually fixed.